prepare for production w/ tsdx
