+++ .gitignore
@@ -1,4 +1,6 @@
-# See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
+*.log
+node_modules
+.cache
 
 # dependencies
 /node_modules
@@ -10,6 +12,7 @@
 
 # production
 /build
+dist
 
 # misc
 .DS_Store
