--- .npmignore
@@ -1,26 +1,0 @@
-webpack.local.config.js
-    webpack.production.config.js
-    .eslintrc
-    .gitignore
-
-# dependencies
-/node_modules
-/.pnp
-.pnp.js
-
-# testing
-/coverage
-
-# production
-/build
-
-# misc
-.DS_Store
-.env.local
-.env.development.local
-.env.test.local
-.env.production.local
-
-npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
