--- .vscode/settings.json
@@ -1,10 +1,0 @@
-{
-    "cSpell.words": [
-        "FBAN",
-        "FBAV",
-        "Twilio",
-        "UUID's",
-        "detectrtc",
-        "toastify"
-    ]
-}
