+++ src/components/index.tsx
@@ -1,5 +1,5 @@
-import HeaderComponent from "./header_comp";
-import IncompatibleComponent from "./incomp_comp";
-import ChatComponent from "./chat_comp";
+import HeaderComponent from './header_comp';
+import IncompatibleComponent from './incomp_comp';
+import ChatComponent from './chat_comp';
 
 export { HeaderComponent, IncompatibleComponent, ChatComponent };
