+++ src/typings/audio.d.ts
