+++ src/typings/interfaces.ts
